// frontend/generated/no/sivertsensoftware/contactregistration/model/Contact.ts
interface Contact {
    id?: number;
    firstname?: string;
    lastname?: string;
    companyname?: string;
    address?: string;
    city?: string;
    county?: string;
    state?: string;
    zip?: string;
    phonenumber?: string;
    phonenumber2?: string;
    email?: string;
}
export default Contact;